backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.7
requests==2.31.0
stripe==7.7.0
python-dateutil==2.8.2
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
boto3==1.29.7
django-storages==1.14.2
pytz==2023.3
cryptography==41.0.7
pycryptodome==3.19.0
webauthn==1.11.1
qrcode==7.4.2
reportlab==4.0.7
weasyprint==60.2
premailer==3.10.0 